cc3xx: extend psa aead tests to verify inline PSA operations
When PSA_FRAMEWORK_HAS_MM_IOVEC is disabled, input / output buffers are
not directly passed to the backend PSA calls, but scratch buffers are
used.

However, when PSA_FRAMEWORK_HAS_MM_IOVEC is enabled, it forces
zero-copying by passing the buffers directly. As there are cases where
the input buffer is the same as the output buffer, it is important to
cover the testing of such scenarios. This change achieves that.

Change-Id: Ie8d048efa406c700ebd46fd362e123a21111eb91
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>