LIB: Move TEST_LOG to use tfm_log
Move the existing TEST_LOG macro to use the new tfm_log API and remove
tfm_log_raw.c. This further unifies the logging APIs within the various
TF-M components.

For the case of the NS testsuite, we always set the USE_STDIO flag in
CMake which causes all NS test logging to use printf. This is
consistent with the NS image as a whole (which uses printf for its
logging).

Change-Id: If391d80ccb254964a1bf5fc13d98e2093bb2d9c6
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>