Build: Fix collect_build_cmd_args handling when a list is used
The collect_build_cmd_args function failed when command-line
arguments contained a list.
(For example, TFM_EXTRA_MANIFEST_LIST_FILES is a list and can
contain multiple paths when multiple out-of-tree secure partitions
are used.)
This commit fixes the issue by treating parameters as lists and
iterating through them. Each element is normalized individually if it’s
a relative path, and semicolons are escaped to keep the -D argument as
a single list item.

Change-Id: Ieb18b25ba20396572077ef13258d8a7dfe0f2917
Signed-off-by: Bence Balogh <bence.balogh@arm.com>