Crypto: remove testing on null input/output buffers in BL1 hashing
TF-M's PSA thin layer relies on assertion to check buffer validity.
Runtime checks are unnecessary as the buffers are controlled and not fed
externally to BL1, also there is not recovery path thus returning an
error in runtime is futile.

Consequently, remove Crypto tests which expect runtime checks of buffer
validity in BL1.

Change-Id: Ibd7d90bbb1c73e478ebabe29319b6714f8d9cc16
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>